fix int

+++ src/views/PrintOuts/Misc_Items_PDF/Invoice.js
@@ -20,9 +20,9 @@ export default (data, pricing) => {
 
   const prices = data.misc_items.map((i) => {
     if (i.category === 'preselect') {
-      return parseFloat(i.qty) * currency(i.price).value;
+      return parseInt(i.qty) * currency(i.price).value;
     } else {
-      return i.pricePer ? parseFloat(i.qty) * currency(i.pricePer).value : 0;
+      return i.pricePer ? parseInt(i.qty) * currency(i.pricePer).value : 0;
     }
   });
